+++ examples/example-protocol/src/assets/rust_wasmer_runtime_test/expected_bindings.rs
@@ -38,6 +38,190 @@ impl Runtime {
         Store::new(&engine)
     }
 
+    pub fn export_array_f32(&self, arg: [f32; 3]) -> Result<[f32; 3], InvocationError> {
+        let arg = serialize_to_vec(&arg);
+        let result = self.export_array_f32_raw(arg);
+        let result = result.map(|ref data| deserialize_from_slice(data));
+        result
+    }
+    pub fn export_array_f32_raw(&self, arg: Vec<u8>) -> Result<Vec<u8>, InvocationError> {
+        let mut env = RuntimeInstanceData::default();
+        let import_object = create_import_object(self.module.store(), &env);
+        let instance = Instance::new(&self.module, &import_object).unwrap();
+        env.init_with_instance(&instance).unwrap();
+        let arg = export_to_guest_raw(&env, arg);
+        let function = instance
+            .exports
+            .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_array_f32")
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_array_f32".to_owned())
+            })?;
+        let result = function.call(arg.to_abi())?;
+        let result = import_from_guest_raw(&env, result);
+        Ok(result)
+    }
+
+    pub fn export_array_f64(&self, arg: [f64; 3]) -> Result<[f64; 3], InvocationError> {
+        let arg = serialize_to_vec(&arg);
+        let result = self.export_array_f64_raw(arg);
+        let result = result.map(|ref data| deserialize_from_slice(data));
+        result
+    }
+    pub fn export_array_f64_raw(&self, arg: Vec<u8>) -> Result<Vec<u8>, InvocationError> {
+        let mut env = RuntimeInstanceData::default();
+        let import_object = create_import_object(self.module.store(), &env);
+        let instance = Instance::new(&self.module, &import_object).unwrap();
+        env.init_with_instance(&instance).unwrap();
+        let arg = export_to_guest_raw(&env, arg);
+        let function = instance
+            .exports
+            .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_array_f64")
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_array_f64".to_owned())
+            })?;
+        let result = function.call(arg.to_abi())?;
+        let result = import_from_guest_raw(&env, result);
+        Ok(result)
+    }
+
+    pub fn export_array_i16(&self, arg: [i16; 3]) -> Result<[i16; 3], InvocationError> {
+        let arg = serialize_to_vec(&arg);
+        let result = self.export_array_i16_raw(arg);
+        let result = result.map(|ref data| deserialize_from_slice(data));
+        result
+    }
+    pub fn export_array_i16_raw(&self, arg: Vec<u8>) -> Result<Vec<u8>, InvocationError> {
+        let mut env = RuntimeInstanceData::default();
+        let import_object = create_import_object(self.module.store(), &env);
+        let instance = Instance::new(&self.module, &import_object).unwrap();
+        env.init_with_instance(&instance).unwrap();
+        let arg = export_to_guest_raw(&env, arg);
+        let function = instance
+            .exports
+            .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_array_i16")
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_array_i16".to_owned())
+            })?;
+        let result = function.call(arg.to_abi())?;
+        let result = import_from_guest_raw(&env, result);
+        Ok(result)
+    }
+
+    pub fn export_array_i32(&self, arg: [i32; 3]) -> Result<[i32; 3], InvocationError> {
+        let arg = serialize_to_vec(&arg);
+        let result = self.export_array_i32_raw(arg);
+        let result = result.map(|ref data| deserialize_from_slice(data));
+        result
+    }
+    pub fn export_array_i32_raw(&self, arg: Vec<u8>) -> Result<Vec<u8>, InvocationError> {
+        let mut env = RuntimeInstanceData::default();
+        let import_object = create_import_object(self.module.store(), &env);
+        let instance = Instance::new(&self.module, &import_object).unwrap();
+        env.init_with_instance(&instance).unwrap();
+        let arg = export_to_guest_raw(&env, arg);
+        let function = instance
+            .exports
+            .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_array_i32")
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_array_i32".to_owned())
+            })?;
+        let result = function.call(arg.to_abi())?;
+        let result = import_from_guest_raw(&env, result);
+        Ok(result)
+    }
+
+    pub fn export_array_i8(&self, arg: [i8; 3]) -> Result<[i8; 3], InvocationError> {
+        let arg = serialize_to_vec(&arg);
+        let result = self.export_array_i8_raw(arg);
+        let result = result.map(|ref data| deserialize_from_slice(data));
+        result
+    }
+    pub fn export_array_i8_raw(&self, arg: Vec<u8>) -> Result<Vec<u8>, InvocationError> {
+        let mut env = RuntimeInstanceData::default();
+        let import_object = create_import_object(self.module.store(), &env);
+        let instance = Instance::new(&self.module, &import_object).unwrap();
+        env.init_with_instance(&instance).unwrap();
+        let arg = export_to_guest_raw(&env, arg);
+        let function = instance
+            .exports
+            .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_array_i8")
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_array_i8".to_owned())
+            })?;
+        let result = function.call(arg.to_abi())?;
+        let result = import_from_guest_raw(&env, result);
+        Ok(result)
+    }
+
+    pub fn export_array_u16(&self, arg: [u16; 3]) -> Result<[u16; 3], InvocationError> {
+        let arg = serialize_to_vec(&arg);
+        let result = self.export_array_u16_raw(arg);
+        let result = result.map(|ref data| deserialize_from_slice(data));
+        result
+    }
+    pub fn export_array_u16_raw(&self, arg: Vec<u8>) -> Result<Vec<u8>, InvocationError> {
+        let mut env = RuntimeInstanceData::default();
+        let import_object = create_import_object(self.module.store(), &env);
+        let instance = Instance::new(&self.module, &import_object).unwrap();
+        env.init_with_instance(&instance).unwrap();
+        let arg = export_to_guest_raw(&env, arg);
+        let function = instance
+            .exports
+            .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_array_u16")
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_array_u16".to_owned())
+            })?;
+        let result = function.call(arg.to_abi())?;
+        let result = import_from_guest_raw(&env, result);
+        Ok(result)
+    }
+
+    pub fn export_array_u32(&self, arg: [u32; 3]) -> Result<[u32; 3], InvocationError> {
+        let arg = serialize_to_vec(&arg);
+        let result = self.export_array_u32_raw(arg);
+        let result = result.map(|ref data| deserialize_from_slice(data));
+        result
+    }
+    pub fn export_array_u32_raw(&self, arg: Vec<u8>) -> Result<Vec<u8>, InvocationError> {
+        let mut env = RuntimeInstanceData::default();
+        let import_object = create_import_object(self.module.store(), &env);
+        let instance = Instance::new(&self.module, &import_object).unwrap();
+        env.init_with_instance(&instance).unwrap();
+        let arg = export_to_guest_raw(&env, arg);
+        let function = instance
+            .exports
+            .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_array_u32")
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_array_u32".to_owned())
+            })?;
+        let result = function.call(arg.to_abi())?;
+        let result = import_from_guest_raw(&env, result);
+        Ok(result)
+    }
+
+    pub fn export_array_u8(&self, arg: [u8; 3]) -> Result<[u8; 3], InvocationError> {
+        let arg = serialize_to_vec(&arg);
+        let result = self.export_array_u8_raw(arg);
+        let result = result.map(|ref data| deserialize_from_slice(data));
+        result
+    }
+    pub fn export_array_u8_raw(&self, arg: Vec<u8>) -> Result<Vec<u8>, InvocationError> {
+        let mut env = RuntimeInstanceData::default();
+        let import_object = create_import_object(self.module.store(), &env);
+        let instance = Instance::new(&self.module, &import_object).unwrap();
+        env.init_with_instance(&instance).unwrap();
+        let arg = export_to_guest_raw(&env, arg);
+        let function = instance
+            .exports
+            .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_array_u8")
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_array_u8".to_owned())
+            })?;
+        let result = function.call(arg.to_abi())?;
+        let result = import_from_guest_raw(&env, result);
+        Ok(result)
+    }
+
     pub async fn export_async_struct(
         &self,
         arg1: FpPropertyRenaming,
@@ -64,7 +248,9 @@ impl Runtime {
             .get_native_function::<(FatPtr, <u64 as WasmAbi>::AbiType), FatPtr>(
                 "__fp_gen_export_async_struct",
             )
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_async_struct".to_owned())
+            })?;
         let result = function.call(arg1.to_abi(), arg2.to_abi())?;
         let result = ModuleRawFuture::new(env.clone(), result).await;
         Ok(result)
@@ -91,7 +277,11 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_fp_adjacently_tagged")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported(
+                    "__fp_gen_export_fp_adjacently_tagged".to_owned(),
+                )
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -115,7 +305,9 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_fp_enum")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_fp_enum".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -136,7 +328,9 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_fp_flatten")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_fp_flatten".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -163,7 +357,11 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_fp_internally_tagged")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported(
+                    "__fp_gen_export_fp_internally_tagged".to_owned(),
+                )
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -187,7 +385,9 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_fp_struct")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_fp_struct".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -208,7 +408,9 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_fp_untagged")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_fp_untagged".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -232,7 +434,9 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_generics")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_generics".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -251,7 +455,9 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<(), FatPtr>("__fp_gen_export_get_bytes")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_get_bytes".to_owned())
+            })?;
         let result = function.call()?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -302,7 +508,11 @@ impl Runtime {
             .get_native_function::<(<i8 as WasmAbi>::AbiType, FatPtr), <i64 as WasmAbi>::AbiType>(
                 "__fp_gen_export_multiple_primitives",
             )
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported(
+                    "__fp_gen_export_multiple_primitives".to_owned(),
+                )
+            })?;
         let result = function.call(arg1.to_abi(), arg2.to_abi())?;
         let result = WasmAbi::from_abi(result);
         Ok(result)
@@ -322,7 +532,9 @@ impl Runtime {
             .get_native_function::<<bool as WasmAbi>::AbiType, <bool as WasmAbi>::AbiType>(
                 "__fp_gen_export_primitive_bool",
             )
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_primitive_bool".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = WasmAbi::from_abi(result);
         Ok(result)
@@ -342,7 +554,9 @@ impl Runtime {
             .get_native_function::<<f32 as WasmAbi>::AbiType, <f32 as WasmAbi>::AbiType>(
                 "__fp_gen_export_primitive_f32",
             )
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_primitive_f32".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = WasmAbi::from_abi(result);
         Ok(result)
@@ -362,7 +576,9 @@ impl Runtime {
             .get_native_function::<<f64 as WasmAbi>::AbiType, <f64 as WasmAbi>::AbiType>(
                 "__fp_gen_export_primitive_f64",
             )
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_primitive_f64".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = WasmAbi::from_abi(result);
         Ok(result)
@@ -382,7 +598,9 @@ impl Runtime {
             .get_native_function::<<i16 as WasmAbi>::AbiType, <i16 as WasmAbi>::AbiType>(
                 "__fp_gen_export_primitive_i16",
             )
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_primitive_i16".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = WasmAbi::from_abi(result);
         Ok(result)
@@ -402,7 +620,9 @@ impl Runtime {
             .get_native_function::<<i32 as WasmAbi>::AbiType, <i32 as WasmAbi>::AbiType>(
                 "__fp_gen_export_primitive_i32",
             )
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_primitive_i32".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = WasmAbi::from_abi(result);
         Ok(result)
@@ -422,7 +642,9 @@ impl Runtime {
             .get_native_function::<<i64 as WasmAbi>::AbiType, <i64 as WasmAbi>::AbiType>(
                 "__fp_gen_export_primitive_i64",
             )
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_primitive_i64".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = WasmAbi::from_abi(result);
         Ok(result)
@@ -442,7 +664,9 @@ impl Runtime {
             .get_native_function::<<i8 as WasmAbi>::AbiType, <i8 as WasmAbi>::AbiType>(
                 "__fp_gen_export_primitive_i8",
             )
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_primitive_i8".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = WasmAbi::from_abi(result);
         Ok(result)
@@ -462,7 +686,9 @@ impl Runtime {
             .get_native_function::<<u16 as WasmAbi>::AbiType, <u16 as WasmAbi>::AbiType>(
                 "__fp_gen_export_primitive_u16",
             )
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_primitive_u16".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = WasmAbi::from_abi(result);
         Ok(result)
@@ -482,7 +708,9 @@ impl Runtime {
             .get_native_function::<<u32 as WasmAbi>::AbiType, <u32 as WasmAbi>::AbiType>(
                 "__fp_gen_export_primitive_u32",
             )
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_primitive_u32".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = WasmAbi::from_abi(result);
         Ok(result)
@@ -502,7 +730,9 @@ impl Runtime {
             .get_native_function::<<u64 as WasmAbi>::AbiType, <u64 as WasmAbi>::AbiType>(
                 "__fp_gen_export_primitive_u64",
             )
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_primitive_u64".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = WasmAbi::from_abi(result);
         Ok(result)
@@ -522,7 +752,9 @@ impl Runtime {
             .get_native_function::<<u8 as WasmAbi>::AbiType, <u8 as WasmAbi>::AbiType>(
                 "__fp_gen_export_primitive_u8",
             )
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_primitive_u8".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = WasmAbi::from_abi(result);
         Ok(result)
@@ -549,7 +781,11 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_serde_adjacently_tagged")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported(
+                    "__fp_gen_export_serde_adjacently_tagged".to_owned(),
+                )
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -573,7 +809,9 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_serde_enum")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_serde_enum".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -594,7 +832,9 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_serde_flatten")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_serde_flatten".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -621,7 +861,11 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_serde_internally_tagged")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported(
+                    "__fp_gen_export_serde_internally_tagged".to_owned(),
+                )
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -645,7 +889,9 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_serde_struct")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_serde_struct".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -669,7 +915,9 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_serde_untagged")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_serde_untagged".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -690,7 +938,9 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_string")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_string".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -711,7 +961,9 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_export_timestamp")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_timestamp".to_owned())
+            })?;
         let result = function.call(arg.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -729,7 +981,9 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<(), ()>("__fp_gen_export_void_function")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_export_void_function".to_owned())
+            })?;
         let result = function.call()?;
         let result = WasmAbi::from_abi(result);
         Ok(result)
@@ -755,7 +1009,7 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_fetch_data")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| InvocationError::FunctionNotExported("__fp_gen_fetch_data".to_owned()))?;
         let result = function.call(r#type.to_abi())?;
         let result = ModuleRawFuture::new(env.clone(), result).await;
         Ok(result)
@@ -774,7 +1028,7 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<(), ()>("__fp_gen_init")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| InvocationError::FunctionNotExported("__fp_gen_init".to_owned()))?;
         let result = function.call()?;
         let result = WasmAbi::from_abi(result);
         Ok(result)
@@ -796,7 +1050,9 @@ impl Runtime {
         let function = instance
             .exports
             .get_native_function::<FatPtr, FatPtr>("__fp_gen_reducer_bridge")
-            .map_err(|_| InvocationError::FunctionNotExported)?;
+            .map_err(|_| {
+                InvocationError::FunctionNotExported("__fp_gen_reducer_bridge".to_owned())
+            })?;
         let result = function.call(action.to_abi())?;
         let result = import_from_guest_raw(&env, result);
         Ok(result)
@@ -807,6 +1063,14 @@ fn create_import_object(store: &Store, env: &RuntimeInstanceData) -> ImportObjec
     imports! {
         "fp" => {
             "__fp_host_resolve_async_value" => Function::new_native_with_env(store, env.clone(), resolve_async_value),
+            "__fp_gen_import_array_f32" => Function::new_native_with_env(store, env.clone(), _import_array_f32),
+            "__fp_gen_import_array_f64" => Function::new_native_with_env(store, env.clone(), _import_array_f64),
+            "__fp_gen_import_array_i16" => Function::new_native_with_env(store, env.clone(), _import_array_i16),
+            "__fp_gen_import_array_i32" => Function::new_native_with_env(store, env.clone(), _import_array_i32),
+            "__fp_gen_import_array_i8" => Function::new_native_with_env(store, env.clone(), _import_array_i8),
+            "__fp_gen_import_array_u16" => Function::new_native_with_env(store, env.clone(), _import_array_u16),
+            "__fp_gen_import_array_u32" => Function::new_native_with_env(store, env.clone(), _import_array_u32),
+            "__fp_gen_import_array_u8" => Function::new_native_with_env(store, env.clone(), _import_array_u8),
             "__fp_gen_import_explicit_bound_point" => Function::new_native_with_env(store, env.clone(), _import_explicit_bound_point),
             "__fp_gen_import_fp_adjacently_tagged" => Function::new_native_with_env(store, env.clone(), _import_fp_adjacently_tagged),
             "__fp_gen_import_fp_enum" => Function::new_native_with_env(store, env.clone(), _import_fp_enum),
@@ -844,6 +1108,54 @@ fn create_import_object(store: &Store, env: &RuntimeInstanceData) -> ImportObjec
             "__fp_gen_make_http_request" => Function::new_native_with_env(store, env.clone(), _make_http_request),
         }
     }
+}
+
+pub fn _import_array_f32(env: &RuntimeInstanceData, arg: FatPtr) -> FatPtr {
+    let arg = import_from_guest::<[f32; 3]>(env, arg);
+    let result = super::import_array_f32(arg);
+    export_to_guest(env, &result)
+}
+
+pub fn _import_array_f64(env: &RuntimeInstanceData, arg: FatPtr) -> FatPtr {
+    let arg = import_from_guest::<[f64; 3]>(env, arg);
+    let result = super::import_array_f64(arg);
+    export_to_guest(env, &result)
+}
+
+pub fn _import_array_i16(env: &RuntimeInstanceData, arg: FatPtr) -> FatPtr {
+    let arg = import_from_guest::<[i16; 3]>(env, arg);
+    let result = super::import_array_i16(arg);
+    export_to_guest(env, &result)
+}
+
+pub fn _import_array_i32(env: &RuntimeInstanceData, arg: FatPtr) -> FatPtr {
+    let arg = import_from_guest::<[i32; 3]>(env, arg);
+    let result = super::import_array_i32(arg);
+    export_to_guest(env, &result)
+}
+
+pub fn _import_array_i8(env: &RuntimeInstanceData, arg: FatPtr) -> FatPtr {
+    let arg = import_from_guest::<[i8; 3]>(env, arg);
+    let result = super::import_array_i8(arg);
+    export_to_guest(env, &result)
+}
+
+pub fn _import_array_u16(env: &RuntimeInstanceData, arg: FatPtr) -> FatPtr {
+    let arg = import_from_guest::<[u16; 3]>(env, arg);
+    let result = super::import_array_u16(arg);
+    export_to_guest(env, &result)
+}
+
+pub fn _import_array_u32(env: &RuntimeInstanceData, arg: FatPtr) -> FatPtr {
+    let arg = import_from_guest::<[u32; 3]>(env, arg);
+    let result = super::import_array_u32(arg);
+    export_to_guest(env, &result)
+}
+
+pub fn _import_array_u8(env: &RuntimeInstanceData, arg: FatPtr) -> FatPtr {
+    let arg = import_from_guest::<[u8; 3]>(env, arg);
+    let result = super::import_array_u8(arg);
+    export_to_guest(env, &result)
 }
 
 pub fn _import_explicit_bound_point(env: &RuntimeInstanceData, arg: FatPtr) {