+++ macros/src/utils.rs
@@ -1,14 +1,29 @@
 use std::collections::VecDeque;
 
+use crate::CollectableTypeDefinition;
 use proc_macro::TokenStream;
 use proc_macro2::Ident;
+use std::str::FromStr;
 use syn::{
-    punctuated::Punctuated, Generics, Item, ItemUse, Path, PathArguments, PathSegment, ReturnType,
-    Type,
+    punctuated::Punctuated, Expr, ExprLit, Generics, Item, ItemUse, Lit, Path, PathArguments,
+    PathSegment, ReturnType, Type, TypeArray,
 };
 
-pub(crate) fn extract_path_from_type(ty: &Type) -> Option<Path> {
+pub(crate) fn extract_path_from_type(ty: &Type) -> Option<CollectableTypeDefinition> {
     match ty {
+        Type::Array(TypeArray {
+            elem,
+            len: Expr::Lit(ExprLit {
+                lit: Lit::Int(len), ..
+            }),
+            ..
+        }) => {
+            let def = extract_path_from_type(elem).unwrap();
+            Some(CollectableTypeDefinition {
+                path: def.path,
+                array_len: usize::from_str(len.base10_digits()).unwrap(),
+            })
+        }
         Type::Path(path) if path.qself.is_none() => {
             let mut path = path.path.clone();
             for segment in &mut path.segments {
@@ -19,7 +34,7 @@ pub(crate) fn extract_path_from_type(ty: &Type) -> Option<Path> {
                     args.colon2_token = Some(syn::parse_quote!(::));
                 }
             }
-            Some(path)
+            Some(CollectableTypeDefinition { path, array_len: 0 })
         }
         _ => None,
     }